IMPALA-8906: Fix flaky profile observability test
The test_query_profile_contains_query_compilation_metadata_load_events
is flaky because only non-zero stats are printed and the test is not
prepared to handle lines appearing based on conditions.

In general, zero values are printed in the profile as well, this makes
profile parsing simpler. Therefore, instead of changing the test the
condition has been removed as part of this change and the test is
updated accordingly.

Change-Id: I7f6172d75294c1dea8f6be086ebb303725c92620
Reviewed-on: http://gerrit.cloudera.org:8080/14704
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>