IMPALA-9160: Remove references to RangerAuthorizationConfig
Due to recent changes in
https://issues.apache.org/jira/browse/RANGER-2646, the way Impala
instantiates a Ranger authorization configuration in
RangerAuthorizationConfig.java should change accordingly.

Two steps are thus required.

1. Remove the references to the class of RangerConfiguration in
RangerAuthorizationConfig.java.
2. Instantiate rangerConfig_ using the class of RangerPluginConfig
provided in the newer version of Ranger instead of RangerConfiguration
as is done today.

The second step requires us to bump up CDP_BUILD_NUMBER to 1626038 or a
later number. Due to compatibility issues of Impala with other
components in the toolchain, for now we just perform the first step.
The second step will be implemented in a follow-up JIRA after we can be
sure there will not be any compatibility issue after bumping up
CDP_BUILD_NUMBER.

Change-Id: I397c73d83a9c6f8414e83d3a605ab15cb171daa5
Reviewed-on: http://gerrit.cloudera.org:8080/14721
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>