IMPALA-4192: Move static state from ExecNode into a PlanNode
This patch adds a new class called PlanNode which contains a subset
of the static state  of their corresponding ExecNode, of which there
is one instance per fragment. ExecNode contains the runtime state
and there can be up to MT_DOP instances of it per fragment.

It also adds a similar class called AggregatorConfig which contains
static state initialized from the thrift aggregator struct and is
passed as an input to the Aggregator class's constructor.

Eventually all static state including codegened function pointers
would be moved to the PlanNodes.

Testing:
Ran exhaustive tests successfully.

Change-Id: I69f1676bf67bac31fa5902511b3fcc269fd67472
Reviewed-on: http://gerrit.cloudera.org:8080/14764
Reviewed-by: Bikramjeet Vig <bikramjeet.vig@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>