IMPALA-9188: Composite primary keys should use same constraint name
After IMPALA-9104, dataload started failing when USE_CDP_HIVE=true
because of HIVE-16603 (Enforce foreign keys to refer to primary keys or
unique keys), which is in Hive 3 but not Hive 2.

This patch fixes a bug from IMPALA-2112 where Impala was generating a
unique constraint name for each column in a composite primary key,
rather than using the same constraint name for each column.

Testing:
* Ran dataload + core tests with USE_CDP_HIVE=[true/false]

Change-Id: I7a91658945b0355753c3cf05be195757b37edaf4
Reviewed-on: http://gerrit.cloudera.org:8080/14792
Reviewed-by: Sahil Takiar <stakiar@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>