IMPALA-9235: add more per-connection stats to /rpcz
This backports commit 0f6d33b4a29873197952335a5777ccf9163fc307
from Kudu and makes corresponding Impala changes.

It also exposes some additional information returned by KRPC,
including per-connection information for inbound connections
and in-flight RPCs. Some of these are in the JSON only, others
are exposed in the data tables on the debug page.

This adds some Protobuf -> JSON utilities similar to those
used in Kudu, except instead of outputting strings, they
append to rapidjson documents.

Testing:
Added a sanity test to test_web_pages.

Manually checked the /rpcz page when running queries.

Backport notes:
Mostly this was a clean cherry-pick. I omitted the changes
to be/src/kudu/rpc/rpc-test.cc, since those seem to depend
on previous Kudu changes and we don't run that test anyway.
I also omitted the changes to Kudu's server code, which
we don't have copied here.

The Kudu commit message is reproduced here:
=========================================
rpc: add TCP socket statistics to /rpcz

This adds the ability to fetch various bits of socket-level information
for each RPC connection and publish the info into /rpcz. The information
itself is fetched using getsockopt(TCP_INFO) as well as ioctls to check
the current send and receive queue lengths.

This data can help resolve whether a use case is network bound or bound
by the application itself. For example, a high number of retransmitted
packets can indicate that the network path to the receiver is
overloaded.

Eventually we may want to expose some of this information on a per-call
basis. However, doing so is quite tricky, since 'send()' completes when
the data has been placed into the outbound packet queue and doesn't wait
until the data is ACKed. We'd need to defer checking for retransmissions
until all of the data has been ACKed, which is at some indeterminate
point in the future. The very newest kernels allow subscribing to such
notifications (along with lots of interesting stats) but, given none of
that is available in el7, it's probably not worth tackling at this
point.
=========================================

Change-Id: I3696463e22123fe81073af4aa495a96b7d4f7ee2
Reviewed-on: http://gerrit.cloudera.org:8080/14884
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>