IMPALA-4224: part 1: schedule join builds
This adds the scheduler logic for fragments with join builds at their
root. These fragments need to be co-located with the fragment with the
join node.

The new code is not active yet because the planner does not generate
plans with join builds (except for planner tests). This change
was validated in the context of a larger patch that enables the join
build plans via the planner and makes query execution work.

Change-Id: I779463cfa2ea9b372607d2be6d5d2252a6469e34
Reviewed-on: http://gerrit.cloudera.org:8080/14944
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>