IMPALA-8501: Fix race condition on port in RpcMgrTest
RpcMgrTest previously would choose a random, unused port when starting
and then eventually start a krpc service on the port. This resulted in
a race condition where the port could be taken by another process
between when it was chosen and when RpcMgrTest actually used it,
causing RpcMgrTest to fail.

This patch significantly reduces the probability of such an error
happening by choosing the port immediately before it is used.

Change-Id: I841196f8557f29816f5f2960f36fc7772233b975
Reviewed-on: http://gerrit.cloudera.org:8080/14997
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>