IMPALA-9295: Read aux error info regardless of overall query status
IMPALA-9137 added AuxErrorInfoPB to FragmentInstanceExecStatusPB
in order to propagate additional error information from Impala executors
to coordinators. The extra error information helps the coordinator decide
if it should blacklist any nodes.

Prior to this patch, the aux error info was only read if the
overall_status of the ReportExecStatusRequestPB was an error. However,
Impala executors don't synchronize the setting of the the aux error info
and the overall_status, so it is possible the aux error info is dropped.

This patch fixes this issue by always looking for the aux error info
from each fragment, regardless of the overall_status from the report.

Testing:
* Core tests pass

Change-Id: I7dfb5efd7a70626dfc5227f28288ff52ebb4861a
Reviewed-on: http://gerrit.cloudera.org:8080/15027
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>